requirements.txt
rembg==2.0.67
pillow==11.3.0
onnxruntime==1.22.1
opencv-python==4.12.0.88
flask==3.0.0